FEM: geom tools, move get element
