+++ src/Mod/Fem/femtools/geomtools.py
@@ -174,6 +174,24 @@ def is_same_geometry(
 
 
 # ************************************************************************************************
+def get_element(
+    part,
+    element
+):
+    if element.startswith("Solid"):
+        index = int(element.lstrip("Solid")) - 1
+        if index >= len(part.Shape.Solids):
+            FreeCAD.Console.PrintError(
+                "Index out of range. This Solid does not exist in the Shape!\n"
+            )
+            return None
+        else:
+            return part.Shape.Solids[index]  # Solid
+    else:
+        return part.Shape.getElement(element)  # Face, Edge, Vertex
+
+
+# ************************************************************************************************
 def get_rectangular_coords(
     obj
 ):
